Modified Create / Insert methods
Added extra column o_abbr that is used when attempting to create the
games as the website we use does match the abbreviations we use for the
teams.  o_abbr stands for Other Abbreviation.
Removed unused picture column.
Changed the wins / losses from int( 2 ) to int( 11 ) to match other
columns.

Added the _Default_Teams method, which returns an array of all the
teams that need to be inserted.  This private method is called from
the Create method.