Changed the date column to an UNIX timestamp
The date column has been changed to a UNIX timestamp value instead of
a datetime column.  This potenitally allows for additional types of
databases to be used in the future.  In addition, we can now display
timestamps in the locale format of the user using JS.

In addition, added the Create_Weeks() method which creates all 17.
This method should be triggered on site install.