Fixed issue in _Run_Statement where the bind_params variable was incorrect
A passed in variable was bind_params, but we also created a local array for the bind_params
to be passed to mysqli_bind_params and the 2 were stepping all over each other.

In addition, fixed issue where the result set was not being returned correctly in the single
method.