The Database class no longer extends the mysqli class
Instead we save the mysqli connector to the member _mysqli and all references
to the mysqli class are done through that object now.  In addition, added the
new public function insertID() which returns the mysqli->insert_id record.  This
saves the hassle of callers needing to know the Database class object layout.

Updated the 3 places that directly called db->insert_id to now call db->insertID()