The execute method now returns when an error is encountered
Originally we were not doing this because the validate method would
return false on validation errors.  The validate method should now
return true on validation errors and we will check for those errors
being set before calling update now.  This allows the content to still
be rendered on validation errors.