Added concept of weekly records
Weekly records has been something that has been needed for years.  Due to the
fact we have a high volume of users that miss one or more picks multiple weeks
out of the year, it became very difficult to get accurate records for everyone.
This was mainly an issue because of the "hangover" rule where if a user misses
all their games (or most) they would only get 2 more losses than the worst record.
But because of the way the data is stored (no weekly records) it was hard to determine
everyone's record for the week.  What if someone missed all games and another user missed
all but the Thursday night game?  We wouldn't want to give the 1 user 2 worse records than
the other user that only picked 1 game.  We'd want to give them both 2 more losses than
the worst record for the week based on someone that got most of their games in.

Weekly records should allow us to manually track each user on a week by week basis.  In
addition, you can only edit an individual user's record if they have 1 or more missing
games.  When manually updating a weekly record the "manual" flag will be set which allows
us to indiciate at runtime what the record was manually calculated.

There is still some legacy code around the picks.picked column / old style ways of updating
records.  From a quick glance none of those code paths are hit still but I'll dive into it
more after verifying all these chagnes work as expected.