Fixed SQL injection issue in _Load_Users
That was probably the only place in the software that had SQL injection since
we took the sort value and put it directly into the query.  The sort value
is now verified as a valid column.

This wasn't really an issue however since the JSON entrypoint was only accessible
via an admin account anyway.