65: Reworked the leaderboard to look more like an official leaderboard
The leaderboard now outputs a specific wins and losses column making it easier
to read.  In addition, it indicates players that have opted out of the perfect
week pool.  Finally, the players in the top 5 spots (the payout spots) now output
their position using green text to indicate they are in the money.