73: Overhauled database connection design
Created new DadtabaseManager that handles the database connection and
all of the associated tables.  All database tables now derive from an
abstract base class DatabaseTable which allows the individual classes
to not worry about the constructor and setup of the class.  This also
allows all of the screens / JSON entry points to directly access the
table classes instead of requiring them to create a new instance of the
class each time it had to be referenced.

There are still a large number of things that need to be cleaned up,
mainly in the ScreenRendere and JSONManager classes with how all of this
works but that will come in pieces.