73: Converted the DatabaseTableUsers class over to only doing database related functionality
This involved removing the "user" specific members from the class "id", "account", etc...
In addition, the CreateSession has been moved to the Login screen as that is where the
session should be created.  Also the ValidateSession functionality moved into the
Authentication class and now happens as part of the initialization of the class.