Slightly tweaked the unpaid perfect week query
The previous query would only report users that had at least one loss
so if a weekly record had to be manually updated because they forgot to
make their picks, they would not show up until the weekly record was set.

The new query will now everyone that does NOT have the expected number of
wins for the week.