Modified the cookieid / token session generation to use openssl_random_pseudo_bytes
This also fixes the hilarious issue, that one person signing in, caused all
other users to be signed into that account.  This was due to the recent commit
that removed session_start() from index.php.  The previous cookieid value was
generated using session_id() but since a session didn't exist, a blank value
was output, which always caused the same cookieid value to be generated.  Luckily
this wasn't pushed to production.

The new code generates 64 bytes of random data that is then hashed with sha256
to generate a 64 byte length value.  The sessions database table was also increased
to use exactly 64 bytes.