Added the ability to modify a user password
This greatly helps out when someone cannot reset their password for whatever reason.
The new password will be set as a temporary password and the user will be asked to
change it on the next login.

In addition, moved the delete user functionality into its own dialog instead of
using an awful prompt format.