Added additional HTML_ROOT configuration setting
This setting will be used to define the HTML root, which is necessary for
dynamically loading the database table classes when executing cron jobs.
Since cron jobs are not executed from the web root, the glob function call
that finds the PHP files was not finding any files so no databases were
being loaded.