Moved version detection into screen manager
This allows all screens to load the version if necessary.  This was primarily done
so that the value could be used in the cache busting T= parameter for references to
the JS files, so that when a version change occurs browsers will load the new file
from the server instead of loading the data from the browsers cache.